🏗️  Split authenticate.js
Split authenticate.js file into two separate files containing each function from the original file. Also removed unneeded dependencies from modules.
